requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
easyocr
numpy
pillow
requests
